Refactor test module names #48
