+++ src/tests/db/tools_for_set_db_entity_test.py
@@ -3,7 +3,7 @@ from src.db.base_set_db_entity import ISetDbEntity
 from src.schemas.abstract_types import K
 from src.schemas.abstract_types import V
 
-from .test_set_db_classes import SetTestData
+from .classes_for_set_db_test import SetTestData
 
 
 async def check_sets_count(
